+++ visualization/dot-visualizer/src/test/java/net/automatalib/visualization/dot/ProviderTest.java
@@ -21,7 +21,6 @@ import java.util.Random;
 import javax.swing.SwingUtilities;
 
 import net.automatalib.automaton.fsa.impl.CompactDFA;
-import net.automatalib.common.util.system.JVMUtil;
 import net.automatalib.visualization.VPManager;
 import net.automatalib.visualization.Visualization;
 import net.automatalib.visualization.VisualizationProvider;
@@ -58,7 +57,7 @@ public class ProviderTest {
     @Test(dependsOnMethods = "testProviderConfiguration", timeOut = 10000)
     public void testDisplay() throws InterruptedException, InvocationTargetException {
 
-        if (!(JVMUtil.getCanonicalSpecVersion() == 11)) {
+        if (!(Runtime.version().feature() == 11)) {
             throw new SkipException("The headless AWT environment currently only works with Java 11 or <=8");
         }